Import DIAMOND2 output file

When performing BLAST searches with the blast_*() functions, the corresponding BLAST output file can be imported into the current R session using this function. All output formats given by BLAST are supported (see e.g. description for details).

read_diamond(file, out_format, use_arrow_duckdb_connection = FALSE)

Arguments

file

path to BLAST output file.

out_format

a character string specifying the output format of the BLAST output that shall be imported. Available options are:

out_format = "xml" : XML

out_format = "csv" : Comma-separated values

use_arrow_duckdb_connection

shall DIAMOND2 hit output table be transformed to an in-process (big data disk-processing) arrow connection to DuckDB? This is useful when the DIAMOND2 output table to too large to fit into memory. Default is use_arrow_duckdb_connection = FALSE.

See also

Author

Hajk-Georg Drost